% Generated by roxygen2 (4.0.2): do not edit by hand
\name{collapseNode}
\alias{collapseNode}
\title{Collapse a Node in a Tree (ie, create a polytomy)}
\usage{
collapseNode(phy, node)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}

\item{node}{A numerical node in the tree to collapse}
}
\value{
Returns pruned phylogeny
}
\description{
This function collapse a node in a tree to a polytomy
}
\examples{
tree <- rtree(10)
collapseNode(tree, 17)
}

