% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ConvertFastaToPhylip}
\alias{ConvertFastaToPhylip}
\title{Convert a FASTA formatted file to Phylip}
\usage{
ConvertFastaToPhylip(fasta.file, file.name = "phylipFile.phy")
}
\arguments{
\item{fasta.file}{A fasta formatted file}

\item{file.name}{A new file name}
}
\description{
This function will concatenate a fasta file and write it in a phylip
}

