% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DataNameCheck}
\alias{DataNameCheck}
\title{Compare Datasets}
\usage{
DataNameCheck(data1, data2)
}
\arguments{
\item{data1}{A matrix or data.frame with meaningful rownames}

\item{data2}{A matrix or data.frame with meaningful rownames}
}
\value{
Returns list of data that is not in the other set.
}
\description{
This function will compare rownames of two datasets
}

