% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dropTaxaFromData}
\alias{dropTaxaFromData}
\title{drop taxa from a dataset}
\usage{
dropTaxaFromData(char, phy, toFile = TRUE)
}
\arguments{
\item{char}{A matrix or data.frame with meaningful rownames}

\item{phy}{A phylogenetic tree in the class "phylo"}

\item{toFile}{If TRUE, will save csv file to working directory}
}
\value{
Returns pruned dataset
}
\description{
This function will compare rownames of a dataset with tree taxa and drop the non-matches from the dataset.
}

