% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dropTaxaFromPhy}
\alias{dropTaxaFromPhy}
\title{Drop taxa From Phy}
\usage{
dropTaxaFromPhy(phy, char, toFile = TRUE, treeFormat = match.arg(treeFormat,
  choices = c("newick", "nexus"), several.ok = FALSE))
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}

\item{char}{A character matrix with rownames as taxa}

\item{toFile}{If TURE, it will save pruned tree in working directory}

\item{treeFormat}{either "newick" or "nexus" depending on how you want the tree written}
}
\value{
Returns pruned phylogeny
}
\description{
This function will check that taxonomic names in a tree and a dataset and if they do not, it will write the pruned trees to file.
}
\examples{
tree <- rtree(10)
chars <- matrix(runif(18), ncol=2)
rownames(chars) <- paste0("t", 1:9)
dropTaxaFromPhy(tree, chars, toFile=FALSE)
}
\seealso{
\link{matchTreeTaxa} \link{TreeNameCheck}
}

