% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findRepeats}
\alias{findRepeats}
\title{Find Repeats}
\usage{
findRepeats(x)
}
\arguments{
\item{x}{A vector of numeric data}
}
\value{
Returns pruned data1
}
\description{
This function will find where repeats exist in a vector
}
\examples{
v <- c(1,2,3,3,3,4,5)
findRepeats(v)
}

