% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetAncestors}
\alias{GetAncestors}
\title{Get Tree Ancestors}
\usage{
GetAncestors(tree, tip)
}
\arguments{
\item{tree}{A phylogenetic tree in the class "phylo"}

\item{tip}{A numerical node within phy}
}
\value{
Returns a node number
}
\description{
This function will traverse a tree rootward, and return the all ancestors for a node.
}
\examples{
tree <- rtree(10)
GetAncestors(tree, 9)
}
\seealso{
\link{whichNodesArePolytomies} \link{node.offspring}
}

