% Generated by roxygen2 (4.0.2): do not edit by hand
\name{howManyPolytomies}
\alias{howManyPolytomies}
\title{How many Polytomies in a tree}
\usage{
howManyPolytomies(phy)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}
}
\value{
Returns single value for the number of polytomies in a tree
}
\description{
This function will a single value for the number of polytomies in a tree
}
\examples{
tree <- rtree(10)
howManyPolytomies(tree)
}
\seealso{
\link{whichNodesArePolytomies} \link{node.offspring}
}

