% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.polytomy}
\alias{is.polytomy}
\title{Is a Polytomy}
\usage{
is.polytomy(phy, node)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}

\item{node}{A numerical node within phy}
}
\value{
Returns boolean response; TRUE for is a polytomy
}
\description{
This function will return if a node in a tree is a polytomy (has more than two offspring)
}
\examples{
tree <- rtree(10)
is.polytomy(tree, 11)
}
\seealso{
\link{whichNodesArePolytomies} \link{node.offspring}
}

