% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matchData}
\alias{matchData}
\title{Match Datasets}
\usage{
matchData(data1, data2, toFile)
}
\arguments{
\item{data1}{A matrix or data.frame with meaningful rownames}

\item{data2}{A matrix or data.frame with meaningful rownames}

\item{toFile}{If TRUE, then will write a csv file to the WD}
}
\value{
Returns pruned data1
}
\description{
This function will compare rownames of two datasets and write pruned files to the working directory.
}

