% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matchTreeTaxa}
\alias{matchTreeTaxa}
\title{Check taxa match in two trees}
\usage{
matchTreeTaxa(phy1, phy2, toFile = TRUE, treeFormat = match.arg(treeFormat,
  choices = c("newick", "nexus"), several.ok = FALSE))
}
\arguments{
\item{phy1}{A phylogenetic tree in the class "phylo"}

\item{phy2}{A phylogenetic tree in the class "phylo"}

\item{toFile}{If TRUE, it will save pruned tree files in the working directory}

\item{treeFormat}{either "newick" or "nexus" depending on how you want the tree written}
}
\value{
Returns either "OK" if all taxa match, or it returns a list with taxa from each tree that does not match the other.
}
\description{
This function will check that taxonomic names in two trees match and if they do not, it will write the pruned trees to file.
}
\examples{
tree1 <- rtree(10)
tree2 <- rtree(9)
matchTreeTaxa(tree1, tree2, toFile=FALSE)
}
\seealso{
\link{phyDataMatch} \link{TreeNameCheck}
}

