% Generated by roxygen2 (4.0.2): do not edit by hand
\name{node.leaves}
\alias{node.leaves}
\title{Node Leaves}
\usage{
node.leaves(phy, node)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}

\item{node}{A numerical node within phy}
}
\value{
Returns a vector of taxa
}
\description{
This function will return all the tip taxa from a node in the tree
}
\examples{
tree <- rtree(10)
node.leaves(tree, 15)
}
\seealso{
\link{node.offspring}
}

