% Generated by roxygen2 (4.0.2): do not edit by hand
\name{node.offspring}
\alias{node.offspring}
\title{node offspring}
\usage{
node.offspring(phy, node)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}

\item{node}{A numerical node within phy}
}
\value{
Returns offspring nodes from a tree.  If it is bifurcating, then this function will return two nodes.  This function does not return tip values, for that see node.leaves()
}
\description{
Find offspring nodes
}
\examples{
tree <- rtree(10)
node.offspring(tree, 11)
}
\seealso{
\link{node.leaves} \link{collapseNode}
}

