% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optimizePolytomyPruning}
\alias{optimizePolytomyPruning}
\title{Optimize Pruning Out Polytomies}
\usage{
optimizePolytomyPruning(phy, maxTries = 100, stopAtNumberSameHits = 5)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}

\item{maxTries}{A number of tries to beat the best tree (ie, tree with the most taxa)}

\item{stopAtNumberSameHits}{A kill number.  If it hits this value, then it terminates the search.}
}
\value{
Returns bifurcating tree with the most number of taxa
}
\description{
This function will randomly remove taxa and keep the tree with the most taxa.
}
\examples{
tree <- rcoal(10)
tree <- collapseNode(tree, 18)
NotTheBest <- pruneFromPolytomies(tree)
MaybeBetter <- optimizePolytomyPruning(tree)
}
\seealso{
\link{pruneFromPolytomies}
}

