% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parent.node}
\alias{parent.node}
\title{Single Parent Node}
\usage{
parent.node(phy, node)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}

\item{node}{A numerical node within phy}
}
\value{
Returns a single ancestral node
}
\description{
This function will traverse a tree rootward, and return the most recent ancestral node
}
\examples{
tree <- rtree(10)
parent.node(tree, 15)
}
\seealso{
\link{node.offspring} \link{GetAncestors}
}

