% Generated by roxygen2 (4.0.2): do not edit by hand
\name{phyDataMatch}
\alias{phyDataMatch}
\title{Phy-Data Match}
\usage{
phyDataMatch(phy, data, saveMismatches = FALSE)
}
\arguments{
\item{phy}{phylogenetic tree in the class "phylo"}

\item{data}{A matrix or data.frame with rownames as taxa}

\item{saveMismatches}{Boolean; to save the mismatched data to a file}
}
\description{
This function will compare a dataset and a phylogenetic tree and compare taxa.
}

