% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{phyloextra}
\alias{phyloextra}
\alias{phyloextra-package}
\title{phyloextra - Set of tools that I have found useful over the years}
\description{
A set of totally random functions that I have made over the years to help me code.  There is no rhyme or reason and there may be better fnctions out there that do the same thing.
}
\details{
\tabular{ll}{
Package: \tab phyloextra \cr
Type: \tab Package \cr
Version: \tab 1.0 \cr
Date: \tab 2014-10-25 \cr
License: \tab GPL(>= 2) \cr
}
}
\author{
Barb Banbury
}

