% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotPrior}
\alias{plotPrior}
\title{Plot Prior Distribution}
\usage{
plotPrior(priorFn, priorVariables, plotQuants = TRUE, plotLegend = TRUE)
}
\arguments{
\item{priorFn}{Prior shape; either "fixed", "uniform", "normal", "lognormal", "gamma", or "exponential"}

\item{priorVariables}{Variables needed to describe the shape of the distribution.  uniform distributions=c(min, max); normal distributions=c(mean, sd); lognormal distributions=c(mean, sd); gamma distributions=c(shape, scale); exponential distributions=c(rate)}

\item{plotQuants}{If TRUE, plots line segments at the quantiles}

\item{plotLegend}{If TRUE, plots legend box with quantile values}
}
\value{
Returns a distribution plot
}
\description{
This function will plot the distribution of priors
}
\examples{
plotPrior(priorFn="exponential", priorVariables=c(10))
plotPrior("normal", c(1,2))
plotPrior("gamma", c(2, .2), plotQuants=FALSE, plotLegend=FALSE)
}

