% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removePackage}
\alias{removePackage}
\title{Remove Package}
\usage{
removePackage(packageName)
}
\arguments{
\item{packageName}{An installed package}
}
\description{
I was frustrated with the base uninstall.packages function, because I had to specify a lib path.  For this function, I search where the package is stored and delete all copies of it.
}
\examples{
#install.packages("ape")
#removePackage("ape")
}

