% Generated by roxygen2 (4.0.2): do not edit by hand
\name{whichNodesArePolytomies}
\alias{whichNodesArePolytomies}
\title{Find Polytomies}
\usage{
whichNodesArePolytomies(phy)
}
\arguments{
\item{phy}{A phylogenetic tree in the class "phylo"}
}
\value{
Returns a vector of node numbers in a tree that are polytomies
}
\description{
Find all polytomies in a tree
}
\seealso{
\link{is.polytomy} \link{node.offspring}
}

